mi_dashboard/requirements.txt
dash
pandas
mysql-connector-python
plotly
